Downstream opm-common filesystem reorganization
