+++ examples/grdecl2vtu.cpp
@@ -32,8 +32,8 @@
 #include <opm/grid/CpGrid.hpp>
 
 #include <opm/grid/utility/OpmParserIncludes.hpp>
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 
 using namespace Dune;
