+++ opm/grid/cpgrid/CpGridData.hpp
@@ -74,7 +74,7 @@
 #include <opm/grid/utility/platform_dependent/reenable_warnings.h>
 
 #if HAVE_ECL_INPUT
-#include <opm/parser/eclipse/EclipseState/Grid/NNC.hpp>
+#include <opm/input/eclipse/EclipseState/Grid/NNC.hpp>
 #endif
 
 #include <array>
