+++ opm/grid/polyhedralgrid/dgfparser.hh
@@ -23,9 +23,9 @@
 #include <opm/grid/polyhedralgrid/gridfactory.hh>
 
 #if HAVE_ECL_INPUT
-#include <opm/parser/eclipse/Parser/ParseContext.hpp>
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/ParseContext.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 #endif
 
 namespace Dune
