+++ tests/cpgrid/cpgrid_aquifer_test.cpp
@@ -7,8 +7,8 @@
 
 #include <opm/grid/CpGrid.hpp>
 #include <opm/grid/common/CommunicationUtils.hpp>
-#include <opm/parser/eclipse/Parser/ErrorGuard.hpp>
-#include <opm/parser/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Parser/ErrorGuard.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
 #if HAVE_MPI
 class MPIError {
 public:
