+++ tests/cpgrid/grid_nnc.cpp
@@ -27,9 +27,9 @@
 
 #include <dune/common/version.hh>
 #include <dune/grid/common/mcmgmapper.hh>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
 #include <opm/grid/CpGrid.hpp>
 #include <vector>
 #include <utility>
