+++ tests/test_column_extract.cpp
@@ -10,8 +10,8 @@
 
 #if HAVE_ECL_INPUT
 #include <opm/grid/utility/OpmParserIncludes.hpp>
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 #endif
 
 #include <cstddef>
