+++ tests/test_cpgrid.cpp
@@ -21,8 +21,8 @@ using Dune::referenceElement; //grid check assume usage of Dune::Geometry
 #include <opm/grid/utility/platform_dependent/reenable_warnings.h>
 
 #include <opm/grid/utility/OpmParserIncludes.hpp>
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 #include <iostream>
 
 template <class GridView>
